+++ src/components/Select/types.ts
@@ -106,6 +106,7 @@ export interface TMultiSelectPropTypes extends IFormCompProps, TSelectBaseProps 
   labelAddons?: JSX.Element
   className?: string
   isButtonSelect?: boolean
+  dataId?: string
 }
 
 export interface TButtonSelectPropTypes extends IFormCompProps, TSelectBaseProps {
@@ -149,6 +150,8 @@ export interface TSingleSelectPropTypes extends IFormCompProps, TSelectBaseProps
     floatValue: number | undefined
   }) => boolean
   renderOptions?: (props: IRenderOptionItemProps) => JSX.Element
+  defaultValue?: string
+  shouldRecalculateDropdownPosition?: boolean
 }
 
 export type TSelectFooterPropTypes = {