+++ src/hooks/index.ts
@@ -12,3 +12,5 @@ export * from './useHideBodyScroll'
 export * from './useGetHasBottomSpace'
 export * from './useFieldArray'
 export * from './useTabScroll'
+export * from './useRecalculateDropdownPosition'
+