+++ src/stories/Alert.stories.tsx
@@ -2,6 +2,7 @@ import React from 'react'
 import type { StoryFn } from '@storybook/react'
 import { Alert as AlertComp } from '../index'
 import { TAlertProps } from '../components/Alert/types'
+import IconClock from '../components/SVGIcons/IconClock'
 
 export default {
   title: 'Alert',
@@ -25,6 +26,9 @@ export const Alert = Template.bind({})
 Alert.args = {
   type: 'information',
   position: 'sticky',
+  iconProps: {
+    Component: IconClock
+  },
   text: 'Alert line which displays the main function or reason of the alert.',
   subtext: 'Become a legendary UX/UI designer through real world and practical courses.',
   closeIcon: true,